+++ apps/rosen/src/app/api/v1/assets/validations.ts
@@ -33,7 +33,7 @@ const filterParser = new FilterParser({
     },
     offset: {
       min: 0,
-      max: 1000,
+      max: Infinity,
       default: 0,
     },
   },